+++ app/Http/Controllers/Lawyer/ClientController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Inertia\Inertia;
 use App\Models\Client;
 use App\Models\BankAccounts;
+use App\Models\CaseFiles;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Request as FacadesRequest;
 
@@ -14,6 +15,8 @@ class ClientController extends Controller
 {
     public function index(Request $request)
     {
+        $clientAccounts = BankAccounts::where('bank_account_type_id', 1)->get();
+
         $filters = FacadesRequest::all(['search']);
         $clients = Client::filter(FacadesRequest::only('search'))
             ->paginate(25)
@@ -29,6 +32,19 @@ class ClientController extends Controller
         return Inertia::render('Lawyer/Client/Index', [
             'filters' => $filters,
             'clients' => $clients,
+            'clientAccounts' => $clientAccounts,
+        ]);
+    }
+
+    public function view($client_id)
+    {
+
+        $clientProfile = Client::query()
+            ->where('id', 'like', "%{$client_id}%")
+            ->first();;
+
+        return Inertia::render('Lawyer/Client/View', [
+            'clientProfile' => $clientProfile,
         ]);
     }
 
@@ -59,16 +75,18 @@ class ClientController extends Controller
             'created_by' => Auth::id(),
         ]);
 
-        return redirect()->route('lawyer.client.index')->with('message', 'Successfully added new client.');
+        return redirect()->route('lawyer.client.index')->with('successMessage', 'Successfully added new client.');
     }
 
     public function edit($client_id)
     {
+        $clients = BankAccounts::where('bank_account_type_id', 1)->get();
         $clientProfile = Client::query()
             ->where('id', 'like', "%{$client_id}%")
             ->first();
         return Inertia::render('Lawyer/Client/Edit', [
-            'client' => $clientProfile
+            'clientProfile' => $clientProfile,
+            'clients' => $clients
         ]);
     }
 
@@ -89,15 +107,19 @@ class ClientController extends Controller
             'linked_client_account' => $request->linked_client_account,
         ]);
 
-        return redirect()->route('lawyer.client.index')->with('message', 'Successfully updated the client.');
+        return redirect()->route('lawyer.client.index')->with('successMessage', 'Successfully updated the client.');
     }
 
     public function destroy($client_id)
     {
         $client = Client::findOrFail($client_id);
 
-        $client->delete();
-
-        return redirect()->route('lawyer.client.index')->with('message', 'Successfully deleted the client.');
+        $case = CaseFiles::findOrFail($client->id);
+        if ($case == null) {
+            $client->delete();
+            return redirect()->route('lawyer.client.index')->with('successMessage', "Successfully deleted the client. {$case->id}");
+        } else {
+            return redirect()->route('lawyer.client.index')->with('errorMessage', "Client still have case files");
+        }
     }
 }
